+++ packages/shared/src/index.ts
@@ -11,5 +11,6 @@ export * from './media';
 export * from './media-menu';
 export * from './menu-position';
 export * from './members';
+export * from './mobile-config';
 export * from './quotes';
 export * from './unsaved-changes';